requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
praat-parselmouth==0.4.7
numpy==1.26.4
httpx==0.27.0
